refactor: 深さの型をu32からu8に変更し、関連する部分を更新 fix: BestmoveEmitterの送信エラーハンドリングを改善

+++ packages/rust-core/crates/engine-cli/src/engine_adapter/types.rs
@@ -10,7 +10,7 @@ use engine_core::shogi::Move;
 pub struct ExtendedSearchResult {
     pub best_move: String,
     pub ponder_move: Option<String>,
-    pub depth: u32,
+    pub depth: u8,
     pub score: i32,
     pub pv: Vec<Move>,
     pub node_type: NodeType,
